Refactor: atom 파일면 변경
list구분 없애기
